requirements.txt
Django==1.10.3
djangorestframework==3.5.3
django-cors-headers==1.3.1
django-rest-knox==2.2.0
django-webpack-loader==0.4.1
